requirements.txt
grpcio
grpcio-tools
uvicorn
fastapi
piccolo[sqlite]
pydantic-settings
protobuf
gunicorn
cryptography
passlib
PyJWT